% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRelativeRisk.R
\name{calcRelativeRisk}
\alias{calcRelativeRisk}
\title{Calculate relative risk}
\usage{
calcRelativeRisk(cases, exposure, fudging = TRUE, alpha = 0.05)
}
\arguments{
\item{cases}{vector of cases}

\item{exposure}{vectore of exposures}

\item{alpha}{alpha value}

\item{fudigng}{If fudging should be used}
}
\value{
A list of results
}
\description{
This function computes the relative risk and the two-by-two table
}
\examples{
calcRelativeRisk(cases=linelist$ill, linelist$moreno)
}
\keyword{epidemiology}

