% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campy.R
\docType{data}
\name{campy}
\alias{campy}
\title{Campylobacter outbreak Greece}
\format{A data frame with 74 observations with 7 variables
\describe{
  \item{id}{The id of the person}
  \item{case}{if the person is a case or not}
}}
\source{
\url{http://www.ecdc.europa.com}
}
\usage{
campy
}
\description{
A dataset containing the dataset to the case study Campy Greece
}
\keyword{datasets}

