% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiet.R, R/epietr.R
\docType{package}
\name{epiet}
\alias{epiet}
\alias{epiet-package}
\title{epiet: A package for computating infectious disease data.}
\description{
The \strong{epiet} package

The \strong{epiet} package
}
\section{epiet functions}{

\code{\link{generateOutbreak}}:


\code{\link{generateOutbreak}}:
}

