% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutbreak.R
\name{generateOutbreak}
\alias{generateOutbreak}
\title{Generate random data of an outbreak}
\usage{
generateOutbreak(n = 100, percentageCases = 0.5, pathogen = "Salmonella",
  type = "pointsource", startdate = "2013-09-23", enddate = "2013-10-04",
  dataquality = "messy")
}
\arguments{
\item{n}{Number of persons}

\item{percentageCases}{Number between 0 and 1 representing the percentage of cases}

\item{pathogen}{Quoted character value of the pathogen causing the outbreak. Possible values are: "Salmonella", "Ebola", "Norovirus", "Measles", "Influenza"}

\item{startdate}{Quoted character value of the first possible date of onset of disease with the format "YYYY-MM-DD". For example: "2013-09-23"}

\item{dataquality}{Quoted character value of the output quality of the dataset. Possible values are: "good", "standard", "messy". "standard" introduces Na Values, "messy" introduces NA and other random values.}

\item{tpye}{Quoted character value of type of outbreak. Possible values are: "pointsource", "propageted", "continuing"}

\item{startdate}{Quoted character value of the last possible date of onset of disease with the format "YYYY-MM-DD". For example: "2013-09-23"}
}
\value{
A dataset of a random outbreak
}
\description{
A function to create a sample dataset of an outbreak
}
\keyword{epidemiology}

