% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir}
\alias{ir}
\title{Calculate table for incidence data}
\usage{
ir(cases, exposure, time, output = "plain", digits = 2)
}
\arguments{
\item{cases}{string of case variable}

\item{exposure}{string of exposure variable}

\item{time}{time of exposure for subjects}

\item{output}{type of table output ( e.g. "plain", "rmarkdown")}

\item{data}{data frame we are working on}
}
\value{
tables of results and statistics
   table contains table of cases, time at risk and incidence rate
   split along category exposed/unexposed.
   If option output = "rmarkdown", then no return but markdown
   tables are printed.
}
\description{
This function gives a table of results for incidence data
}
\keyword{epidemiology}

