% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitable.R
\name{multitable}
\alias{multitable}
\title{Show tables of multiple models
this is a wrapper function for producing nice
output tables with the stargazer package adapted
 to the needs of an epidemiologist.}
\usage{
multitable(..., padding = "3px")
}
\description{
Show tables of multiple models
this is a wrapper function for producing nice
output tables with the stargazer package adapted
 to the needs of an epidemiologist.
}

