% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpicurve.R
\name{plotEpicurve}
\alias{plotEpicurve}
\title{Create an epicurve plot}
\usage{
plotEpicurve(data, dates, unitOfTimeInDays = 1, unitOfLabel = "1 week",
  xlab = "Date", ylab = "Number of cases", title = NULL, boxes = "auto",
  group = NA, groupHeading = " ", groupColor = c("#345A8A", "#4F81BD",
  "grey"), fillColor = "steelblue")
}
\arguments{
\item{data}{data frame which includes dates vector}

\item{dates}{name of the dates column (string)}

\item{unitOfTimeInDays}{numeric vector giving bin size}

\item{unitOfLabel}{character vector giving label}

\item{xlab}{label of x-axis}

\item{title}{title of plot}

\item{boxes}{indicates whether to show boxes for cases}

\item{group}{name of the group column or NULL if no grouping}

\item{groupHeading}{description of grouping for legend}

\item{groupColor}{list of group colors}

\item{fillcolor}{color of fill}
}
\description{
A function to create an epicurve according to EPIET standards
}
\keyword{epidemiology}

