% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratios_adapted.R
\name{ratios_adapted}
\alias{ratios_adapted}
\title{calculate risk ratio or odds ratio}
\usage{
ratios_adapted(data, outcome = "outcome", exposures = c("exposure1",
  "exposure2"), measure = "rr", verbose = F)
}
\arguments{
\item{data}{data.frame containing outcome and exposure data}

\item{outcome}{a character vector containing the name of the outcome column of (1s and 0s) in data}

\item{measure}{a character vector indicating either rr (risk ratio) or 'or' (odds ratio)}

\item{verbose}{a logical vector, if TRUE gives full results if FALSE gives restricted results}

\item{exposure}{a character vector containing the names of exposure columns  of (1s and 0s) in data or vectors of exposures (1s and 0s)}

\item{output}{a data.frame containing results for each exposure}
}
\value{
A list of results
}
\description{
Function to create a data.frame of risk ratio or odds ratio results from a 2x2 contingency table for 
multiple exposures. Note: this function gives the equivalent of the cs command in STATA. Date: 08.12.2014
}
\examples{
# define dummy data

df = generateOutbreak(dataquality="good", numberExposures = 10, probabilityOfTrueExposue = 0.9)
# function examples
# function examples
ratios(df, outcome = "ill", exposures = c(names(df)[13:dim(df)[2]]))
}
\author{
Daniel Gardiner, \email{daniel.gardiner@phe.gov.uk}
}
\keyword{epidemiology}

