% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir.R
\name{ir}
\alias{ir}
\title{Calculate table for incidence data}
\usage{
ir(cases, exposure, time, output = "plain")
}
\arguments{
\item{cases}{string of case variable}

\item{exposure}{string of exposure variable}

\item{time}{time at risk for subjects}

\item{output}{type of table output ( e.g. "plain", "rmarkdown")}

\item{data}{data frame we are working on}
}
\value{
a list of tables of results and statistics
   \itemize{
   \item table: contains table of cases, time at risk and incidence
         split along category exposed/unexposed.
   \item stats: contains difference of ir between exposed/unexposed,
   ir ratios between exposed/unexposed, attributable fraction exposed
   and attributable fraction of population.
   }
   If option output = "rmarkdown" is chosen, then no return but markdown
   tables are printed. The default output="plain" returns two matrices with
   the results for further processing.
}
\description{
This function gives a table of results for incidence data.
The results are comparable to the ir function in Stata.
}
\keyword{epidemiology,}
\keyword{incidence}
\keyword{rate}

