% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpicurve.R
\name{plotEpicurve}
\alias{plotEpicurve}
\title{Create an epicurve plot}
\usage{
plotEpicurve(data, dates, unitOfTimeInDays = 1, unitOfLabel = "1 week",
  xTitle = "Date", yTitle = "Number of cases", plotTitle = "none",
  boxes = "auto", group = "none", groupHeading = " ",
  groupColor = c("#345A8A", "#4F81BD", "grey"), fillColor = "steelblue")
}
\arguments{
\item{data}{data frame which includes dates vector}

\item{dates}{name of the dates column (string)}

\item{unitOfTimeInDays}{numeric vector giving bin size}

\item{unitOfLabel}{character vector giving label}

\item{xTitle}{title of x-axis}

\item{yTitle}{title of y-axis}

\item{plotTitle}{title of plot}

\item{boxes}{indicates whether to show boxes for cases}

\item{group}{vector of group membership or "none". Groups are
shown in different colors.}

\item{groupColor}{list of group colors}

\item{fillcolor}{color of fill}
}
\description{
A function to create an epicurve according to EPIET standards
}
\examples{
plotEpicurve()

}
\keyword{epidemiology}

