% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEpicurve.R
\name{plotEpicurve}
\alias{plotEpicurve}
\title{Create an epicurve plot}
\usage{
plotEpicurve(dates = "Date", data = Linelist, unitOfTimeInDays = 1,
  unitOfLabel = "1 week", xTitle = "Date", yTitle = "Number of cases",
  plotTitle = "none", boxes = "auto", group = "none",
  groupHeading = " ", groupColor = c("#345A8A", "#4F81BD", "grey"),
  fillColor = "steelblue")
}
\arguments{
\item{dates}{date vector}

\item{unitOfTimeInDays}{numeric vector giving bin size}

\item{unitOfLabel}{character vector giving label}
}
\description{
A function to create an epicurve according to EPIET standards
}
\examples{
plotEpicurve()

}
\keyword{epidemiology}

