% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiline_excel.R
\name{read_excel_multiline}
\alias{read_excel_multiline}
\title{Read excel datasets with column names in multiple lines}
\usage{
read_excel_multiline(filename, row_collapse = 1, ...)
}
\arguments{
\item{filename}{excel filename}

\item{row_collapse}{number of rows to collapse}

\item{...}{additional parameter(s) passed on to readxl::read_excel}
}
\value{
tibble with the column names from multiple lines
}
\description{
read multiline excel files
}
\details{
This function reads multiline Excel files using the solution proposed by brianwdavis.
}
\examples{

\dontrun{

# example ---------------------------------------------------

#library(readxl)

#file_path <- "https://github.com/brianwdavis/public/raw/master/example.xlsx"
#download.file(file_path, "example.xlsx", mode = "wb")
#read_excel_multiline("example.xlsx", row_collapse = 2)

}

}
\references{
See the solution proposed by brianwdavis: https://github.com/tidyverse/readxl/issues/486#issuecomment-398224438
}
\seealso{
\code{\link[readxl]{read_excel}}
}
